requirements.txt
pymongo[srv]==4.2.0
pandas
numpy
scikit-learn
python-dotenv
PyYAML
dill

-e .  # editable mode
